Creating an independent fixture generation (and parser) which is separate from the integration tests
