+++ Cargo.toml
@@ -5,5 +5,6 @@ members = [
     "akd_mysql",
     "poc",
     "integration_tests",
-    "akd_client"
+    "akd_client",
+    "akd_test_tools"
 ]
