+++ akd_test_tools/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "akd_test_tools"
+version = "0.6.1"
+authors = ["Evan Au <evanau@fb.com>", "Sean Lawlor <seanlawlor@fb.com>"]
+description = "Test generation and loading tooling"
+license = "MIT OR Apache-2.0"
+edition = "2018"
+keywords = ["key-transparency", "akd"]
+repository = "https://github.com/novifinancial/akd"
+readme = "../README.md"
+
+[dependencies]
+winter-crypto = "0.2"
+winter-math = "0.2"
+clap = { version="3", features = ["derive"]}
+tokio = { version = "1.10", features = ["full"] }
+regex = "1.5"
+rand = "0.7"
+serde_yaml = "0.8"
+serde = "1.0"
+async-trait = "0.1"
+
+akd = { path = "../akd", features = ["serde_serialization"], version = "^0.6.1" }
+
+[dev-dependencies]
+assert_fs="1"
+
+akd = { path = "../akd", features = ["public-tests", "rand", "serde_serialization"], version = "^0.6.1" }
