+++ akd_test_tools/src/main.rs
@@ -4,10 +4,10 @@
 // LICENSE-MIT file in the root directory of this source tree and the Apache
 // License, Version 2.0 found in the LICENSE-APACHE file in the root directory
 // of this source tree.
-#[cfg(test)]
-mod test_util;
-#[cfg(test)]
-mod mysql_tests;
 
-#[cfg(test)]
-mod memory_tests;
+use akd_test_tools::fixture_generator;
+
+#[tokio::main]
+async fn main() {
+    fixture_generator::run().await;
+}
