+++ integration_tests/Cargo.toml
@@ -8,15 +8,6 @@ publish = false
 [dependencies]
 winter-crypto = "0.2"
 winter-math = "0.2"
-clap = { version="3", features = ["derive"]}
-tokio = { version = "1.10", features = ["full"] }
-regex = "1.5"
-rand = "0.7"
-serde_yaml = "0.8"
-serde = "1.0"
-async-trait = "0.1"
-
-akd = { path = "../akd", features = ["public-tests", "rand", "serde_serialization"] }
 
 [dev-dependencies]
 log = { version = "0.4.8", features = ["kv_unstable"] }
@@ -26,6 +17,13 @@ once_cell = "1"
 thread-id = "3"
 multi_log = "0.1"
 hex = "0.4.3"
-assert_fs="1"
 
+tokio = { version = "1.10", features = ["full"] }
+# regex = "1.5"
+rand = "0.7"
+# serde_yaml = "0.8"
+# serde = "1.0"
+# async-trait = "0.1"
+
+akd = { path = "../akd", features = ["public-tests", "rand", "serde_serialization"] }
 akd_mysql = { path = "../akd_mysql" }
